session-params-state moved to dashboard module
