+++ src/app/pages/dashboard/state/session-params/session-params-state.interface.ts
@@ -1,5 +1,5 @@
-import { SessionParams } from '../../global/types/session-params.interface';
-import { BackendErrorResponse } from '../../global/types/backend-error-response.interface';
+import { SessionParams } from '../../../../global/types/session-params.interface';
+import { BackendErrorResponse } from '../../../../global/types/backend-error-response.interface';
 
 export interface SessionParamsState {
   readonly isLoading: boolean;
