+++ src/app/pages/dashboard/state/session-params/session-params.actions.ts
@@ -1,6 +1,6 @@
 import { createActionGroup, emptyProps, props } from '@ngrx/store';
-import { SessionParams } from '../../global/types/session-params.interface';
-import { BackendErrorResponse } from '../../global/types/backend-error-response.interface';
+import { SessionParams } from '../../../../global/types/session-params.interface';
+import { BackendErrorResponse } from '../../../../global/types/backend-error-response.interface';
 
 export const sessionParamsActions = createActionGroup({
   source: 'Session Params',
